TAP version 14 - First Draft
This is a first draft at a specification that seeks to ratify existing
behavior of TAP harnesses and producers.

- YAML blocks standardized to 2 space indentation
- Subtests specified to behavior of `Test::More` and `node-tap`.
- Normative advice regarding exit code for harness programs
- Examples and usage comments made language-agnostic.
- Clarification of whitespace and hyphens in test lines.
- Clarification of handling of incorrect lines.
- Specification of Pragma lines

It'd be very helpful for implementors to point to a body of
language-agnostic tests for compliant parsers.  I've got a pretty nice
start at this over at
<https://github.com/substack/tap-parser/tree/master/test/fixtures>, but
we may want to bikeshed the event/property names a bit, and ideally I'd
like to get at least one other implementation passing those tests before
we sign off on them.

Note that this does not go as far as a lot of people would probably like
to see in a forward-looking TAP specification.  No fancy new magic is
added.  However, before we start talking about brand new features, it
seems wise to ratify the features we are already using.

With this change, node-tap and Test::More should be able to simply
change their version number from 13 to 14 in order to be fully
compliant.  Feedback from other TAP producers and harnesses is
necessary before adopting this officially.