Add links to each of the languages on the producers page.
This branch attempts to clean up some of the producers page.
It provides a description for what TAP producers are.

Additionally, it adds internal anchors that can be used to
link directly to a language. These links would be useful for sharing
out. I believe this fixes #9.

There are some comments in #9 that mention proposed solutions being
non-standard Markdown. Gruber specifically mentions using HTML when
Markdown does not have an answer.

> For any markup that is not covered by Markdown’s syntax, you simply
> use HTML itself. There’s no need to preface it or delimit it to
> indicate that you’re switching from Markdown to HTML; you just use the
> tags.

I also fixed the C# header. It seems that Jekyll strips out
trailing # marks in headers.