Update license information for tap specifications
The TAP specifications state that they are licensed under the "same
terms as Perl itself" and then provide a link which is out of date. This
change updates the link to a valid copy of the Perl artistic license,
adds SPDX identifiers at the top of the specifications, and adds a copy
of the Perl artistic license to the repository.

Why add a copy of the license and SPDX identifiers? It turns out that
there are multiple revisions of the Artistic License 1.0; the copy of
the license and the SPDX identifiers make it explicit which version of
the Artistic License 1.0 is used.

Signed-off-by: Brendan Higgins <brendanhiggins@google.com>

Close: #138