tappy is a standalone viewer; move it to the top of the list.
It's nice that there are all kinds of format transducers and libraries
and things, but a TAP newbie's first deployment of a consumer is going to
be as a simple filter through which he views test results.

Therefore tappy and other simple standalone viewers adapted for this
use case should be listed first, probably in their own section.
Unfortunately tappy is at thus point the only consumer I know
about.  It will take a topic expert to do that edit.